Basically all the MAESPA versions that I use or used
# MAESPA-versions
A repository containing all the version of MAESPA
which are used on the project

1- MAESPA_V7, the one used by
F.Charbonnier.
It was used with the Medlyn (2012) conductance model

2-
Maespa_V8.exe, modified by M.Christina & G. Le Maire in order to
implement a new Foliar Temperature calcul routine. Is basically the
version I used with the Tuzet conductance model (2003) on the Aquiares
plot (with the 6 monitored reprouts)

3- MAESPA_VR_TH_TRANS, wich is
basically the same version than Maespa_V8.exe, modified in order to
output transmittance on chosen points (if you set points.dat and
ipoints=1 on the control namelist of the confile)

4- MAESPA_Remko_VR,
is the web version (also called V2 or MAESPA64), downloaded the
08/06/2015. It was modified in oreder to:
- Simulate more than 3
species (4 actualy)
- Simulate a total max of test points equals to the
max gridpoint of the total number of coffees (39273)
- Debug the test
points routine.
I used it to simulate the tranmittance on the ENSAYO
trial plot.