Fix G02TABLE and G12TABLE problem
There was a confusion between G0TABLE, G0TABLEI and G0TABLESPEC.
I removed G02TABLE and replaced it by G02TABLEI (in the
maindeclaration.f90), as it is used in inout.f90.
I replaced G02TABLE in the inout.f90 by G02TABLESPEC also.
Idem for the G12TABLE